requirements.txt
Flask==0.12
Pillow==4.1.1
